добавлена инструкция в приложение, обновлены файл сборки, скрипт inno setup

+++ code/rebuild_all.py
@@ -6,18 +6,8 @@ import PyInstaller.__main__
 
 from service_scripts.clear_dir import remove_pycaches
 from installation_controller import VERSION_APP
+from service_scripts.convert_instruction import convert_docx_to_html
 
-
-def build_python_app():
-    PyInstaller.__main__.run([
-    'installation_controller.py',
-    '--noconsole', 
-    '--hidden-import=pyvisa_py', 
-    '--exclude-module=PyQt6, PySide6, PySide2, pytest, OpenGL',  
-    '--icon=picture/picture/cat.ico',
-    '--clean',
-    '--noconfirm'
-    ])
 
 def build_python_app():
     PyInstaller.__main__.run([
@@ -25,7 +15,7 @@ def build_python_app():
         '--noconsole', 
         '--hidden-import=pyvisa_py', 
         '--hidden-import=multiprocessing',
-        '--exclude-module=PyQt6,PySide6,PySide2',  
+        '--exclude-module=PyQt6,PySide6,PySide2, mammoth, pytest, OpenGL',  
         '--icon=picture/picture/cat.ico',
         '--clean',
         '--noconfirm',
@@ -82,28 +72,48 @@ def build_inno_setup(script_path):
     except FileNotFoundError as e:
         print(e)
 
+def build_instruction():
+    # Определяем базовую директорию (на уровень выше скрипта)
+    base_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+    
+    # Пути по умолчанию
+    default_input = os.path.join(base_dir, "instruction.docx")
+    default_output = os.path.dirname(os.path.abspath(__file__))
+
+    if not os.path.exists(default_input):
+        print(f"Ошибка: Файл инструкции {default_input} не найден")
+        return False
+        
+    html_path = convert_docx_to_html(default_input, default_output)
+
+    return html_path
+
 if __name__ == "__main__":
+    instruction_path = build_instruction()
+    if instruction_path:
 
-    if os.path.exists('dist'):
-        shutil.rmtree('dist')
-
-    build_python_app()
-    
-    if os.path.exists('build'):
-        shutil.rmtree('build')
-
-    if os.path.exists("installation_controller.spec"):
-        os.remove("installation_controller.spec")
-
-    inno_script_path = 'default.iss'
-    build_inno_setup(inno_script_path)
-
-    if os.path.exists('dist'):
-        try:
+        if os.path.exists('dist'):
             shutil.rmtree('dist')
-        except Exception as e:
-            print(f"Error deleting dist folder: {e}")
-    
-    move_file_up_one_directory('installation_controller.exe')
 
-    remove_pycaches(os.getcwd())
+        build_python_app()
+        
+        if os.path.exists('build'):
+            shutil.rmtree('build')
+
+        if os.path.exists("installation_controller.spec"):
+            os.remove("installation_controller.spec")
+
+        inno_script_path = 'default.iss'
+        build_inno_setup(inno_script_path)
+
+        if os.path.exists('dist'):
+            try:
+                shutil.rmtree('dist')
+            except Exception as e:
+                print(f"Error deleting dist folder: {e}")
+        
+        move_file_up_one_directory('installation_controller.exe')
+
+        remove_pycaches(os.getcwd())
+
+        os.remove(instruction_path)
